Fixes bug that prevented constructing reverse index correctly
This makes the redo slice allocate for the case when we want to
construct a reverse index in the original level loop.

Since we are reusing the redo set as the next level's keys set,
we can not overwrite the same underlying array by adding to the
redo set (from index 0 as we progess through the keys) in the
second iteration through the keys set. This is because when
creating the reverse index, we need a third iteration over the
keys set, and if we have added something to the redo set, we have
(most likely) changed the keys set in the second iteration which
will then be different in the third iteration.

Doing this:

bb.redo = bb.redo[:0]

works in the second iteration, since we only overwrite entries in
the underlying array that has already been processed as part of the
keys set.

Hence, the solution here is to allocate a new redo slice for each
new level if we are create a reverse index; otherwise we do the
trick above to avoid the allocation.