Build the reverse index in a third per-level loop
This creates a reverse index in the per-level loop. This adds a
third iteration through the keys. I think this is technically
O(n^3), but since the keys set is reduced for each level, it should
be faster than doing the third iteration (for all keys) after the
construction of the BBHash level vectors. It seems difficult to
avoid the third iteration with the current design.