Removed useless bit vector Set() call in second loop
This simplifies the second loop; we don't actually need to set
the current level's bit vector positions again in the second
loop since we don't reset the bit vector (as was done in the
original implementation).

This also fixes the code comments to more clearly explain each
loop and separate their functionalities.