Simplified sz and wds variables and avoid lvlData-based method
This avoids the lvlData struct and related methods, calling stuff
directly on the lvlVector instead. This exposes more details in
the computeParallel() method, but I think it makes things more
clear. Moreover, we also avoid updating wds and n multiple times,
since we now use sz instead of n, and wds is updated at the bottom
when advancing to the next level, instead of reading again at the
top of the lvl loop. These are minor tweaks that matter little,
except to make things more clear and avoid duplicate work.