New attempt to parallelize construction using partitioned keys
This version partitions the input keys using a simple mod operator
on the key. Note that this assumes that the keys are uniformly
random. If they are not, the partitioning step should be changed
to hash the inputs. We do not do this, since for our usecase the
keys are already random, and so the partitions are expected to be
approximately equal-sized.

This version is significantly faster the previous attempts.