Remove the reverse mapping functionality
The idea for doing a reverse mapping during bbhash construction
didn't pan out as initially hoped; that it would avoid having to
call the Find(k) function for each key to reverse map. However,
since the reverse mapping need to have the entire bbhash completed
there was nothing to be gained from doing this as part of a custom
function in this library. Moreover, the use case we have in mind
for this reverse map is different from what this provides. We want
to map from a bbhash values to peer addresses or other identifiers.