Fixed wire-bits calculation and bits per key accordingly
We previously accounted for the rank vector when computing the
bits per key; since we don't marshal the rank vector isn't marshaled
it shouldn't be counted when computing the bits per key. However,
since we instead need to count the length of the marshaled vectors,
this mistake did not matter. However, we didn't count the 64 bit
header needed to marshal the bit vectors (the number of bit vectors
in the marshaled []byte).

We could technically use fewer bits for these length values, since
we almost never have more than 20 levels (bit vectors in the bbash).
Similarly, the bit vectors also are never going to be 2^64 long;
It would probably suffice to use 32 bits for the bit vectors.