Added New(g, part, keys) and simplified BenchmarkBBHash{New,Find}
This simplifies BenchmarkBBHash{New,Find} to use the new constructor
bbhash.New(gamma, partitions, keys). This allows us to remove the
two loops and instead use partitions = 1 to run the NewSequential()
version. This is likely to be slightly slower than calling the
NewSequential() directly, but seems like a worthwhile compromise
to simplify the benchmark setup. The other drawback is that it
allocates an additional BBHash2{} to wrap the BBHash into.

Moreover, we no longer benchmark the original Parallel variant.