refactor: moved the reverse map into the BBHash struct
Having an empty slice in the BBHash struct isn't so costly in terms
of memory, and the API for the computeWithKeymap() method becomes
identical to the vanila compute() method. Moreover, this also adds
the Key(index) method to return a single key for a given index.
We currently don't expose the full slice, but if that's necessary
for some use case, we may consider adding an iterator or exposing
it more directly.