feat: add -update flag to automatically update golden test file
This updates the bbhash_reproducible_bit_vectors_test.go file with
golden bit vector values; this is useful if we need to change the
hash function or other parts that affect the bit vectors for the
different levels. Previous version required several manual steps,
so this is a minor simplficiation in that we don't have to manually
copy the bit vector and that we can support multiple bit vector
sizes. That is, we need to tweak the wantBitVectors helper to add
more sizes, so it still isn't entirely seamless to add more sizes.