refactor: reproducibility test to support partition bbhash
This fixes the TestReproducibleBitVectors to allow use of multiple
partitions with bbhash. The BBHash2 is not compatible with the old
BBHash since the keys get partitioned with BBHash2; therefore, this
refactor was quite extensive and the golden file needs to have a
separate entry for each partition size. The Sequential and Parallel
variants are still BBHash compatible.