fix(BBHash): bug in BBHash.marshaledLength; add test for empty BBHash
This fixes a bug in BBHash's marshaledLength() method; the bug was that we only included the number of words required by a bitVector, and missed the fact that a bitVector must also include the header field to specify how many words are included in a bitVector. This fix calls bitVector.marshaledLength() instead of using the bitVector.words() method.

This also makes a few minor tweaks in BBHash.AppendBinary and UnmarshalBinary.

Adds TestMarshalUnmarshalBBHashEmpty.